LPS-72263 rename
