+++ modules/apps/portal-mobile-device-detection-fiftyonedegrees/app.bnd
