+++ modules/apps/portal-mobile-device-detection-fiftyonedegrees/build.gradle
