+++ modules/apps/portal-mobile-device-detection-fiftyonedegrees/portal-mobile-device-detection-fiftyonedegrees/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Portal Mobile Device Recognition Provider FiftyOneDegrees
-Bundle-SymbolicName: com.liferay.portal.mobile.device.detection.fiftyonedegrees.lite
+Bundle-Name: Liferay Portal Mobile Device Detection FiftyOneDegrees
+Bundle-SymbolicName: com.liferay.portal.mobile.device.detection.fiftyonedegrees
 Bundle-Version: 1.0.2
 Export-Package: com.liferay.portal.mobile.device.detection.fiftyonedegrees.data
 Liferay-Releng-Module-Group-Description:
