+++ modules/apps/portal-mobile-device-detection-fiftyonedegrees/portal-mobile-device-detection-fiftyonedegrees/src/main/resources/com/liferay/portal/mobile/device/recognition/provider/fiftyonedegrees/data/packageinfo
