+++ modules/apps/portal-mobile-device-detection-fiftyonedegrees/portal-mobile-device-detection-fiftyonedegrees/src/main/resources/content/Language_uk.properties
